# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentZT Gateway Management API
  description: |
    AgentZT Gateway 管理API提供完整的应用管理、配置管理、统计分析、配额管理和告警管理功能。
    
    ## 认证
    
    所有管理API需要Bearer Token认证。在请求头中添加：
    
    ```
    Authorization: Bearer <your-session-token>
    ```
    
    或使用API Key：
    
    ```
    X-AgentZT-API-Key: <your-api-key>
    ```
    
    ## 应用选择
    
    多应用场景下，可通过以下方式指定应用：
    
    - 请求头：`X-AgentZT-App-ID: <appId>`
    - 请求头：`X-AgentZT-API-Key: <apiKey>`
    - 查询参数：`?appId=<appId>`
    
    ## 错误处理
    
    所有错误响应遵循统一格式：
    
    ```json
    {
      "error": {
        "code": "error_code",
        "message": "Error description",
        "details": {}
      }
    }
    ```
    
    ## 版本
    
    当前API版本为v1，默认路径 `/api` 等同于 `/api/v1`。
  version: 1.0.0
  contact:
    name: AgentZT Team
    email: support@agentzt.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发服务器
  - url: https://gateway.agentzt.io/api
    description: 生产服务器
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session Token认证
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-AgentZT-API-Key
      description: API Key认证
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - authentication_error
                - permission_error
                - not_found
                - invalid_request
                - validation_error
                - conflict_error
                - rate_limit_exceeded
                - internal_error
                - service_unavailable
              example: invalid_request
            message:
              type: string
              example: App name must be between 1 and 100 characters
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              example: req-1699887600000-abc123
            timestamp:
              type: string
              format: date-time
              example: '2024-01-15T10:30:00Z'
...
    AppId:
      type: string
      pattern: '^app-[0-9]+-[a-z0-9]+$'
      example: app-1699887600000-abc123
...
    QuotaType:
      type: string
      enum:
        - checks
        - tokens
        - agents
    QuotaUsage:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/QuotaType'
        used:
          type: integer
        limit:
          type: integer
        percentage:
          type: number
          format: float
        remaining:
          type: integer
    QuotaUsageResponse:
      type: object
      properties:
        appId:
          $ref: '#/components/schemas/AppId'
        checks:
          $ref: '#/components/schemas/QuotaUsage'
        tokens:
          $ref: '#/components/schemas/QuotaUsage'
        alerts:
          type: object
          properties:
            checks:
              type: array
              items:
                type: object
                properties:
                  threshold:
                    type: integer
                  triggered:
                    type: boolean
            tokens:
              type: array
              items:
                type: object
                properties:
                  threshold:
                    type: integer
                  triggered:
                    type: boolean
    QuotaHistoryEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        appId:
          $ref: '#/components/schemas/AppId'
        type:
          $ref: '#/components/schemas/QuotaType'
        delta:
          type: integer
        totalUsed:
          type: integer
        limit:
          type: integer
        requestId:
          type: string
        resource:
          type: string
    QuotaHistoryResponse:
      type: object
      properties:
        appId:
          $ref: '#/components/schemas/AppId'
        timeRange:
          type: string
          enum:
            - day
            - week
            - month
        entries:
          type: array
          items:
            $ref: '#/components/schemas/QuotaHistoryEntry'
        aggregated:
          type: object
          properties:
            checks:
              type: object
              properties:
                total:
                  type: integer
                count:
                  type: integer
                avgDelta:
                  type: number
            tokens:
              type: object
              properties:
                total:
                  type: integer
                count:
                  type: integer
                avgDelta:
                  type: number

    # ========== Alert Schemas ==========
...
  parameters:
...
    AppIdQueryParam:
      name: appId
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/AppId'
      description: 应用ID（可选）
...
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: 返回数量限制
paths:
...
  /quota/usage:
    get:
      tags:
        - Quota
      summary: 获取配额使用情况
      description: 返回当前应用的配额使用情况
      parameters:
        - $ref: '#/components/parameters/AppIdQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaUsageResponse'
        '400':
          description: 缺少appId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quota/history:
    get:
      tags:
        - Quota
      summary: 获取配额使用历史
      description: 返回指定时间范围内的配额使用历史
      parameters:
        - $ref: '#/components/parameters/AppIdQueryParam'
        - name: timeRange
          in: query
          required: false
          schema:
            type: string
            enum:
              - day
              - week
              - month
            default: month
          description: 时间范围
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - checks
              - tokens
          description: 配额类型
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaHistoryResponse'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quota/reset:
    post:
      tags:
        - Quota
      summary: 重置配额使用
      description: 重置指定应用的配额使用（仅管理员）
      parameters:
        - $ref: '#/components/parameters/AppIdQueryParam'
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Quota usage reset for application "app-xxx"
                  appId:
                    $ref: '#/components/schemas/AppId'
        '400':
          description: 缺少appId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限（需要管理员权限）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quota/limit:
    put:
      tags:
        - Quota
      summary: 设置配额限制
      description: 设置指定应用的配额限制（仅管理员）
      parameters:
        - $ref: '#/components/parameters/AppIdQueryParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                checksLimit:
                  type: integer
                  minimum: 0
                  example: 10000
                tokensLimit:
                  type: integer
                  minimum: 0
                  example: 1000000
            example:
              checksLimit: 20000
              tokensLimit: 2000000
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Quota limits updated for application "app-xxx"
                  appId:
                    $ref: '#/components/schemas/AppId'
                  quota:
                    type: object
                    properties:
                      checksLimit:
                        type: integer
                      tokensLimit:
                        type: integer
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限（需要管理员权限）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quota/alerts:
    get:
      tags:
        - Quota
      summary: 获取配额告警历史
      description: 返回指定应用的配额告警历史
      parameters:
        - $ref: '#/components/parameters/AppIdQueryParam'
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - checks
              - tokens
          description: 配额类型
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  appId:
                    $ref: '#/components/schemas/AppId'
                  alerts:
                    type: array
                    items:
                      type: object
                      properties:
                        appId:
                          $ref: '#/components/schemas/AppId'
                        type:
                          $ref: '#/components/schemas/QuotaType'
                        threshold:
                          type: integer
                        percentage:
                          type: number
                          format: float
                        used:
                          type: integer
                        limit:
                          type: integer
                        triggeredAt:
                          type: string
                          format: date-time
                        message:
                          type: string
                  total:
                    type: integer
        '400':
          description: 缺少appId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== Alerts API ==========
...